created reusable badge components
